feat: add displayName to anonymous components
In the React Dev Tools [1] component names are displayed based on either
the class' name, the function's name or a `displayName` property on the
component. When a component is not created as a class or a named
function the component has no display name. To show Chakra UI's
component names correctly in the developer tools this commit introduces
display names to all components that are created using `forwardRef` or
`styled`.

One last problem remains: The some of the `AlertOverlay*` components
are simply reexports from the `Modal` component. When using these
components the name will be wrong in the developer tools. This could be
solved by wrapping with a function that simply forwards props and gets
the name right, but it might not be that much of a big deal so I have
not made this change.

[1]: https://github.com/facebook/react/tree/master/packages/react-devtools