fix(slider): avoid excess onChange calls
This fixes a regression of #2210. It looks like the previous fix was
reverted in fa8273bb41ad667a113b99ed97d1ab39e5790e63 due to the fix
introducing a new issue:

- Begin drag from value X
- Drag to value X + 1 without finishing drag
- Continue drag back to value X is not allowed

That issue was because `onMouseDown`/`onTouchStart` had stale closures
around `value`. Using an evergreen ref of `value` resolves this.